+++ tests/test_json_cache.py
@@ -3,14 +3,14 @@
 import os
 from unittest import mock
 
-from query_and_cache.cacheclass import Cache
-from query_and_cache.json_cache import JSONCache
+from pyquaca.cache import Cache
+from pyquaca.json_cache import JSONCache
 
 
 def test_json_cache_can_instantiate() -> None:
     """Test that the JSONCache class can be instantiated."""
 
-    with mock.patch("query_and_cache.cacheclass.Path") as mock_path:
+    with mock.patch("pyquaca.cache.Path") as mock_path:
         cache = JSONCache("testcache")
         assert mock_path.called
     assert isinstance(cache, JSONCache)
@@ -24,7 +24,7 @@ def test_json_cache_can_instantiate() -> None:
 def test_json_cache_retrieve(mock_file) -> None:
     """Test the JSONCache.retrieve method."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     assert cache.retrieve("test") == {"keystr": "valuestr"}
     cache_path = os.path.join("testcache", "test" + cache.file_extension)
@@ -36,7 +36,7 @@ def test_json_cache_retrieve(mock_file) -> None:
 def test_json_cache_retrieve_file_not_found(mock_file) -> None:
     """Test the JSONCache.retrieve method when file not found."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     mock_file.side_effect = FileNotFoundError
     assert cache.retrieve("test") is None
@@ -49,7 +49,7 @@ def test_json_cache_retrieve_file_not_found(mock_file) -> None:
 def test_json_cache_retrieve_permission_error(mock_file) -> None:
     """Test the JSONCache.retrieve method when permission error."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     mock_file.side_effect = PermissionError
     assert cache.retrieve("test") is None
@@ -64,7 +64,7 @@ def test_json_cache_retrieve_permission_error(mock_file) -> None:
 def test_json_cache_retrieve_io_error(mock_file) -> None:
     """Test the JSONCache.retrieve method when IO error."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     mock_file.side_effect = IOError
     assert cache.retrieve("test") is None
@@ -78,7 +78,7 @@ def test_json_cache_retrieve_io_error(mock_file) -> None:
 def test_json_cache_retrieve_empty_key() -> None:
     """Test the JSONCache.retrieve method when key is empty."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     assert cache.retrieve("") is None
 
@@ -88,7 +88,7 @@ def test_json_cache_retrieve_empty_key() -> None:
 def test_json_cache_store(mock_file) -> None:
     """Test the JSONCache.store method."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     sample_data = {"keystr": "valuestr"}
     assert cache.store("test", sample_data)
@@ -103,7 +103,7 @@ def test_json_cache_store(mock_file) -> None:
 def test_json_cache_store_permission_error(mock_file) -> None:
     """Test the JSONCache.store method when permission error."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     mock_file.side_effect = PermissionError
     assert not cache.store("test", "dummy test text")
@@ -118,7 +118,7 @@ def test_json_cache_store_permission_error(mock_file) -> None:
 def test_json_cache_store_io_error(mock_file) -> None:
     """Test the JSONCache.store method when IO error."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     mock_file.side_effect = IOError
     assert not cache.store("test", {"key": "value"})
@@ -132,7 +132,7 @@ def test_json_cache_store_io_error(mock_file) -> None:
 def test_json_cache_store_empty_key() -> None:
     """Test the JSONCache.store method when key is empty."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     assert not cache.store("", "dummy test text")
 
@@ -141,18 +141,18 @@ def test_json_cache_store_empty_key() -> None:
 def test_json_cache_store_empty_value() -> None:
     """Test the JSONCache.store method when value is empty."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     assert not cache.store("test", "")
 
 
 # Test JSONCache.store returns False when value is invalid
-@mock.patch("query_and_cache.json_cache.json.dumps", side_effect=TypeError)
+@mock.patch("pyquaca.json_cache.json.dumps", side_effect=TypeError)
 @mock.patch("builtins.open", new_callable=mock.mock_open)
 def test_json_cache_store_invalid_value(mock_file, mock_dump) -> None:
     """Test the JSONCache.store method when value is invalid."""
 
-    with mock.patch("query_and_cache.cacheclass.Path"):
+    with mock.patch("pyquaca.cache.Path"):
         cache = JSONCache("testcache")
     assert not cache.store("test", {"key": "value"})
     assert mock_dump.called